build: Get install script under 1 second
The slowest remaining task was 'rename' which seems to be running Perl under the hood? It was taking ~4 seconds to rename 59 files and directories.

Also only create smaller leader images if they don't already exist. This is pretty trivial and only takes a fraction of a second but every little bit helps.